feat: USCC token contract + tests + deploy script (#49)
* feat: USCC token contract

* comment

* max suggestion

* remove unneeded old code

* max comments

* max comments2

+++ test/AllowListV2.sol
@@ -30,7 +30,7 @@ contract AllowListV2 {
     /// @notice A record of permissions for each entityId determining if they are allowed. One indexed, since 0 is the default value for all addresses
     mapping(uint256 => Permission) public permissions;
 
-    /// @notice A record of entityIds associated with each address. Setting to 0 removes the address from the allowList. 
+    /// @notice A record of entityIds associated with each address. Setting to 0 removes the address from the allowList.
     mapping(address => uint256) public addressEntityIds;
 
     /// @notice An event emitted when an entityId's permission status is changed
@@ -90,7 +90,6 @@ contract AllowListV2 {
         emit PermissionSet(entityId, perms);
     }
 
-
     /**
      * @notice Sets the nth permission for a given entityId
      * @param entityId The entityId to be updated
@@ -114,7 +113,11 @@ contract AllowListV2 {
      * @param index The index of the permission to update
      * @param value The status to set
      */
-    function setPermissionAtIndex(Permission memory perms, uint index, bool value) internal pure returns (Permission memory) {
+    function setPermissionAtIndex(Permission memory perms, uint256 index, bool value)
+        internal
+        pure
+        returns (Permission memory)
+    {
         if (index == 0) {
             perms.isAllowed = value;
         } else if (index == 1) {
